// src/types.ts
export type Data<T extends string> = Record<T, string | boolean | undefined>

// Data model RSI uses for spreadsheet imports
export type Fields<T extends string> = DeepReadonly<Field<T>[]>

export interface Field<T extends string> {
  // UI-facing field label
  label: string
  // Field's unique identifier
  key: T
  // UI-facing additional information displayed via tooltip and ? icon
  description?: string
  // Alternate labels used for fields' auto-matching, e.g. "fname" -> "firstName"
  alternateMatches?: string[]
  // Validations used for field entries
  validations?: Validation[]
  // Field entry component, default: Input
  fieldType: Checkbox | Select | Input
  // UI-facing values shown to user as field examples pre-upload phase
  example?: string
}

export interface Checkbox {
  type: "checkbox"
  // Alternate values to be treated as booleans, e.g. {yes: true, no: false}
  booleanMatches?: Record<string, boolean>
}

export interface Select {
  type: "select"
  // Options displayed in Select component
  options: SelectOption[]
}

export interface SelectOption {
  // UI-facing option label
  label: string
  // Field entry matching criteria as well as select output
  value: string
}

export interface Input {
  type: "input"
}

export type Validation = RequiredValidation | UniqueValidation | RegexValidation